refactor(ux): combined sources to one section
- Renamed "Documents" section to "Sources" in the dashboard layout.
- Updated routing for adding sources and managing documents.
- Refactored the connectors and documents upload pages to redirect to the new sources section.
- Added localization support for the new "Sources" terminology in English and Chinese.
