+++ surfsense_web/components/sidebar/nav-main.tsx
@@ -36,13 +36,9 @@ export function NavMain({ items }: { items: NavItem[] }) {
 		const titleMap: Record<string, string> = {
 			Researcher: "researcher",
 			"Manage LLMs": "manage_llms",
-			Documents: "documents",
-			"Upload Documents": "upload_documents",
-			"Add Webpages": "add_webpages",
-			"Add Youtube Videos": "add_youtube",
+			Sources: "sources",
+			"Add Sources": "add_sources",
 			"Manage Documents": "manage_documents",
-			Connectors: "connectors",
-			"Add Connector": "add_connector",
 			"Manage Connectors": "manage_connectors",
 			Podcasts: "podcasts",
 			Logs: "logs",
