+++ surfsense_web/components/sources/types.ts
@@ -1,0 +1,13 @@
+export interface Connector {
+	id: string;
+	title: string;
+	description: string;
+	icon: React.ReactNode;
+	status: "available" | "coming-soon" | "connected";
+}
+
+export interface ConnectorCategory {
+	id: string;
+	title: string;
+	connectors: Connector[];
+}
